// src/common/enums/returncode.enums.ts
enum ReturnCodes {
	SUCCESS = 200,
	CREATE_SUCCESS = 201,
	UPDATE_SUCCESS = 202,
	READ_SUCCESS = 203,
	DELETE_SUCCESS = 204,
	NO_RECORDS_FOUND = 205,

	BAD_REQUEST = 400,
	UNAUTHORIZED_REQUEST = 401,
	RESTRICTED_REQUEST = 412,
	FORBIDDEN_REQUEST = 403,
	UNKNOWN_REQUEST = 404,
	PRECONDITION_FAILED = 412,
	UNPROCESSED_REQUEST = 413,

	ERROR = 500,
}

export default ReturnCodes